+++ plastid/__init__.py
@@ -30,5 +30,5 @@ plastid is divided into the following subpackages:
     ==============    =========================================================
      
 """
-__version__ = "0.2.2"
+__version__ = "0.2.3"
 __author__  = "Joshua Griffin Dunn"