+++ CHANGES.rst
@@ -23,6 +23,7 @@ the last digit.
     Anyway, we don't plan on changing the name again anytime soon. Please pardon
     the confusion!
 
+
 Unreleased
 ----------
 
@@ -47,6 +48,16 @@ Removed
   - sort_segments_lexically and sort_segmentchains_lexically
     removed, because GenomicSegments and SegmentChains now sort
     lexically without help
+
+
+plastid [0.2.3] = [2015-09-23]
+------------------------------
+
+Changed
+.......
+  - Cython implementations of BAM mapping rules now default,
+    are 2-10x faster than Python implementations
+
 
 plastid [0.2.2] = [2015-09-15]
 ------------------------------